// src/frontend/generated/UpdatePendingOperation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UpdateChannel } from "./UpdateChannel";
import type { UpdateOperationState } from "./UpdateOperationState";

export type UpdatePendingOperation = { operationId: string, correlationId: string, channel: UpdateChannel, componentId: string, fromVersion: string, toVersion: string, state: UpdateOperationState, startedAtMs: number, };